add UI test for integer validators for less and greater
