+++ test_suite/tests/ui/integer/validate/bounds/greater_and_greater_or_equal.rs
@@ -1,0 +1,6 @@
+use nutype::nutype;
+
+#[nutype(validate(greater = -272, greater_or_equal = -272))]
+pub struct Amount(i32);
+
+fn main() {}
