+++ test_suite/tests/ui/integer/validate/bounds/less_and_less_or_equal.rs
@@ -1,0 +1,6 @@
+use nutype::nutype;
+
+#[nutype(validate(less = 99, less_or_equal = 99))]
+pub struct Amount(i16);
+
+fn main() {}
