+++ test_suite/tests/ui/integer/validate/bounds/upper_vs_lower.rs
@@ -1,0 +1,6 @@
+use nutype::nutype;
+
+#[nutype(validate(less_or_equal = 0, greater = 20))]
+pub struct Amount(i64);
+
+fn main() {}
