--- test_suite/tests/ui/integer/validate/greater_or_equal_vs_less_or_equal.rs
@@ -1,6 +1,0 @@
-use nutype::nutype;
-
-#[nutype(validate(less_or_equal = 0, greater_or_equal = 20))]
-pub struct Amount(i64);
-
-fn main() {}
